fastapi/requirements.txt
fastapi
uvicorn
motor
pydantic
python-decouple
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
datetime